+++ packages/host-api/src/protocol/v1/remotePermission.ts
@@ -1,6 +1,7 @@
 import { Enum } from '@novasamatech/scale';
-import { Result, Vector, _void, bool, str } from 'scale-ts';
+import { Vector, _void, bool, str } from 'scale-ts';
 
+import { CallResult } from '../callError.js';
 import { GenericError } from '../commonCodecs.js';
 
 export const RemotePermission = Enum({
@@ -12,4 +13,4 @@ export const RemotePermission = Enum({
 });
 
 export const RemotePermissionV1_request = RemotePermission;
-export const RemotePermissionV1_response = Result(bool, GenericError);
+export const RemotePermissionV1_response = CallResult(bool, GenericError);